pyproject.toml
[project]
name = "f1-regulations-agent-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "langchain",
    "langchain-google-genai",
    "langchain-community",
    "faiss-cpu",
    "python-decouple>=3.8",
    "pymupdf>=1.27.2.2",
    "rich>=14.3.3",
    "google-cloud-secret-manager>=2.26.0",
    "google-cloud-aiplatform>=1.141.0",
    "vertexai>=1.43.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "rank-bm25>=0.2.2",
    "google-cloud-vectorsearch>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["f1_agent"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
